Move milestones progress bars partial to a component
