+++ app/assets/stylesheets/application.scss
@@ -27,6 +27,7 @@
 @import "layout/**/*";
 @import "machine_learning/**/*";
 @import "management/**/*";
+@import "milestones/**/*";
 @import "moderation/**/*";
 @import "polls/**/*";
 @import "proposals/**/*";
