--- app/views/milestones/_progress_bars.html.erb
@@ -1,24 +1,0 @@
-<% if milestoneable.primary_progress_bar %>
-  <section class="progress-bars">
-    <h5><%= t("milestones.index.progress") %></h5>
-
-    <div class="margin">
-      <%= progress_tag_for(milestoneable.primary_progress_bar) %>
-    </div>
-
-    <% if milestoneable.secondary_progress_bars.any? %>
-      <div class="milestone-progress">
-        <% milestoneable.secondary_progress_bars.each do |progress_bar| %>
-          <div class="row margin-bottom">
-            <div class="small-12 medium-6 large-4 column">
-              <%= progress_bar.title %>
-            </div>
-            <div class="small-12 medium-6 large-8 column end">
-              <%= progress_tag_for(progress_bar) %>
-            </div>
-          </div>
-        <% end %>
-      </div>
-    <% end %>
-  </section>
-<% end %>
